Replace tests with examples as to not break CI builds.
