+++ src/gitlab.rs
@@ -123,42 +123,9 @@ impl GitLab {
 /*
 #[cfg(test)]
 mod tests {
-    use std::env;
-    use gitlab::GitLab;
-    use gitlab::Pagination;
 
     #[test]
-    fn list_groups() {
-        let token = match env::var("GITLAB_TOKEN") {
-            Ok(val) => val,
-            Err(_)  => panic!("Please set environment variable 'GITLAB_TOKEN'"),
-        };
-
-        let mut gl = GitLab::new_https("gitlab.com", &token);
-        gl.set_pagination(Pagination{page: 1, per_page: 100});
-        println!("gl: {:?}", gl);
-
-        let groups = gl.groups().unwrap();
-        println!("groups: {:?}", groups);
-    }
-
-    #[test]
-    fn list_projects() {
-        let token = match env::var("GITLAB_TOKEN") {
-            Ok(val) => val,
-            Err(_)  => panic!("Please set environment variable 'GITLAB_TOKEN'"),
-        };
-
-        let mut gl = GitLab::new_https("gitlab.com", &token);
-        // for i in 1..82 {
-        //     gl.set_pagination(Pagination{page: i, per_page: 1});
-        //     println!("projects: {:?}", gl.projects().unwrap());
-        // }
-        gl.set_pagination(Pagination{page: 1, per_page: 100});
-        let projects = gl.projects().unwrap();
-        for project in projects {
-            println!("{:?}", project.path_with_namespace);
-        }
+    fn it_works() {
     }
 }
 */
